# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_email.eventListItem:
    properties:
      created_at:
        example: "2026-06-10T12:00:00Z"
        type: string
      id:
        example: evt_001
        type: string
      sender:
        example: partner@company.com
        type: string
      source:
        example: email
        type: string
      status:
        example: completed
        type: string
      subject:
        example: Contract Review Meeting
        type: string
    type: object
  internal_gmail.WatchRequest:
    properties:
      email_address:
        example: user@company.com
        format: email
        type: string
    type: object
  internal_gmail.WatchResponse:
    properties:
      email_address:
        example: user@company.com
        type: string
      expires_at:
        example: "2026-06-17T15:26:00Z"
        type: string
      status:
        example: watch_started
        type: string
    type: object
  internal_gmail.pubSubPushBody:
    properties:
      message:
        properties:
          attributes:
            additionalProperties:
              type: string
            type: object
          data:
            type: string
          messageId:
            type: string
        type: object
      subscription:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@mengu.ai
    name: API Support
  description: AI-assisted email automation platform. Ingests emails, analyzes them
    with LLM to extract intent and actions, then executes actions through deterministic
    handlers.
  license:
    name: Proprietary
  termsOfService: https://mengu.ai/terms
  title: Mengu AI API
  version: 1.0.0
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: List incoming events for the authenticated user's organization.
      parameters:
      - description: Filter by status (new, processing, completed, failed)
        in: query
        name: status
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_email.eventListItem'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: List events
      tags:
      - Events
  /events/{id}:
    get:
      description: Get a single event by ID (without analysis/logs nesting — use GET
        /events/:id for the full detail view).
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              metadata:
                type: object
              source:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get event
      tags:
      - Events
...
  /events/{id}/reanalyze:
    post:
      consumes:
      - application/json
      description: Reset event status to 'new' so the worker re-processes it through
        AI analysis and action execution.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Re-analyze event
      tags:
      - Events
  /gmail/watch:
    post:
      consumes:
      - application/json
      description: Initiates Gmail API watch for the authenticated user's organization
        — subscribes to mailbox push notifications (requires prior Google OAuth consent).
      parameters:
      - description: Email address to watch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gmail.WatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_gmail.WatchResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Start Gmail watch
      tags:
      - Gmail
  /organization:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's organization details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              created_at:
                type: string
              id:
                type: string
              name:
                type: string
              plan:
                type: string
              slug:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get organization
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      description: Update organization settings (name, plan).
      parameters:
      - description: Organization update fields
        in: body
        name: request
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              created_at:
                type: string
              id:
                type: string
              name:
                type: string
              plan:
                type: string
              slug:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Update organization
      tags:
      - Organization
...
  /webhooks/gmail:
    post:
      consumes:
      - application/json
      description: 'Receives Pub/Sub push notifications from Google for Gmail mailbox
        changes (new email, label changes). Idempotent: responds 200 immediately after
        logging and updates historyId.'
      parameters:
      - description: Pub/Sub push notification envelope
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gmail.pubSubPushBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Gmail push notification
      tags:
      - Gmail
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  WebhookSecret:
    description: Webhook secret key for authenticating email webhook requests.
    in: header
    name: X-Webhook-Secret
    type: apiKey
swagger: "2.0"